formatted + updated tests.

+++ contracts/Migrations.sol
@@ -1,15 +1,16 @@
 // SPDX-License-Identifier: Unlicensed
 pragma solidity >=0.6.2 <0.8.0;
+
 contract Migrations {
-  address public owner = msg.sender;
-  uint public last_completed_migration;
+    address public owner = msg.sender;
+    uint256 public last_completed_migration;
 
-  modifier restricted() {
-    require(msg.sender == owner);
-    _;
-  }
+    modifier restricted() {
+        require(msg.sender == owner);
+        _;
+    }
 
-  function setCompleted(uint completed) public restricted {
-    last_completed_migration = completed;
-  }
+    function setCompleted(uint256 completed) public restricted {
+        last_completed_migration = completed;
+    }
 }
