+++ contracts/RFT.sol
@@ -1,16 +1,16 @@
 // SPDX-License-Identifier: Unlicensed
 pragma solidity >=0.6.2 <0.8.0;
 
-import '@openzeppelin/contracts/token/ERC721/IERC721.sol';
-import '@openzeppelin/contracts/token/ERC20/IERC20.sol';
-import '@openzeppelin/contracts/token/ERC20/ERC20.sol';
+import "@openzeppelin/contracts/token/ERC721/IERC721.sol";
+import "@openzeppelin/contracts/token/ERC20/IERC20.sol";
+import "@openzeppelin/contracts/token/ERC20/ERC20.sol";
 
-contract RFT is ERC20{
-    uint public icoSharePrice;
-    uint public icoShareSupply;
-    uint public icoEnd;
+contract RFT is ERC20 {
+    uint256 public icoSharePrice;
+    uint256 public icoShareSupply;
+    uint256 public icoEnd;
 
-    uint public nftId;
+    uint256 public nftId;
     IERC721 public nft;
     IERC20 public dai;
 
@@ -20,12 +20,11 @@ contract RFT is ERC20{
         string memory _name,
         string memory _symbol,
         address _nftAddress,
-        uint _nftId,
-        uint _icoSharePrice,
-        uint _icoShareSupply,
+        uint256 _nftId,
+        uint256 _icoSharePrice,
+        uint256 _icoShareSupply,
         address _daiAddress
-    ) ERC20(_name, _symbol)
-    {
+    ) ERC20(_name, _symbol) {
         nftId = _nftId;
         nft = IERC721(_nftAddress);
         icoSharePrice = _icoSharePrice;
@@ -34,19 +33,21 @@ contract RFT is ERC20{
         admin = msg.sender;
     }
 
-    function startIco()external {
+    function startIco() external {
         require(msg.sender == admin);
         nft.transferFrom(msg.sender, address(this), nftId);
         icoEnd = block.timestamp + 7 * 86400;
-
     }
 
-    function buyShare(uint shareAmount) external {
-        require(icoEnd>0, "ico not started yet.");
+    function buyShare(uint256 shareAmount) external {
+        require(icoEnd > 0, "ico not started yet.");
         require(block.timestamp <= icoEnd, "ico is finished.");
-        require(totalSupply()+shareAmount <= icoShareSupply, "Not enough shares left.");
+        require(
+            totalSupply() + shareAmount <= icoShareSupply,
+            "Not enough shares left."
+        );
 
-        uint daiAmount = shareAmount*icoSharePrice;
+        uint256 daiAmount = shareAmount * icoSharePrice;
         dai.transferFrom(msg.sender, address(this), daiAmount);
         _mint(msg.sender, shareAmount);
     }
@@ -55,12 +56,12 @@ contract RFT is ERC20{
         require(msg.sender == admin, "admin only.");
         require(block.timestamp > icoEnd, "ico not finished yet.");
 
-        uint daiBalance = dai.balanceOf(address(this));
-        if(daiBalance > 0){
+        uint256 daiBalance = dai.balanceOf(address(this));
+        if (daiBalance > 0) {
             dai.transfer(admin, daiBalance);
         }
-        uint unsoldShareBalance = icoShareSupply - totalSupply();
-        if(unsoldShareBalance>0){
+        uint256 unsoldShareBalance = icoShareSupply - totalSupply();
+        if (unsoldShareBalance > 0) {
             _mint(admin, unsoldShareBalance);
         }
     }
